fix dtype mismatch in expand_op test (#1283)
* expand.py

* int64

* docs updated

* test data updated

+++ onnx/backend/test/case/node/expand.py
@@ -20,11 +20,11 @@ class Expand(Base):
             outputs=['expanded'],
         )
         shape = [3, 1]
-        data = np.reshape(np.arange(1, np.prod(shape) + 1, dtype=np.float64), shape)
+        data = np.reshape(np.arange(1, np.prod(shape) + 1, dtype=np.float32), shape)
         #print(data)
         #[[1.], [2.], [3.]]
         new_shape = [2, 1, 6]
-        expanded = data * np.ones(new_shape)
+        expanded = data * np.ones(new_shape, dtype=np.float32)
         #print(expanded)
         #[[[1., 1., 1., 1., 1., 1.],
         #  [2., 2., 2., 2., 2., 2.],
@@ -33,7 +33,7 @@ class Expand(Base):
         # [[1., 1., 1., 1., 1., 1.],
         #  [2., 2., 2., 2., 2., 2.],
         #  [3., 3., 3., 3., 3., 3.]]]
-        new_shape = np.array(new_shape)
+        new_shape = np.array(new_shape, dtype=np.int64)
         expect(node, inputs=[data, new_shape], outputs=[expanded],
                name='test_expand_dim_changed')
 
@@ -46,7 +46,7 @@ class Expand(Base):
         )
         shape = [3, 1]
         new_shape = [3, 4]
-        data = np.reshape(np.arange(1, np.prod(shape) + 1, dtype=np.float64), shape)
+        data = np.reshape(np.arange(1, np.prod(shape) + 1, dtype=np.float32), shape)
         #print(data)
         #[[1.], [2.], [3.]]
         expanded = np.tile(data, 4)
@@ -54,6 +54,6 @@ class Expand(Base):
         #[[1., 1., 1., 1.],
         # [2., 2., 2., 2.],
         # [3., 3., 3., 3.]]
-        new_shape = np.array(new_shape)
+        new_shape = np.array(new_shape, dtype=np.int64)
         expect(node, inputs=[data, new_shape], outputs=[expanded],
                name='test_expand_dim_unchanged')
